minor refactoring - keep _numba.py as majorly numba only code
